% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatting.R
\name{chatting}
\alias{chatting}
\title{R interface to tencent AI lab's chatting robot}
\usage{
chatting(app_id = app_id, app_key = app_key, keywords)
}
\arguments{
\item{app_id}{Your app ID}

\item{app_key}{Your app key}

\item{text}{The text is sent to API}
}
\value{
returns the reply of chatting robots
}
\description{
\code{chatting} returns the reply of chatting robot
}
\examples{
 chatting('hello', app_id = '1107152791', app_key =  'OpsMj8HXPmbu9SMd')
}
\author{
Ao Sun <\url{https://ao-sun.github.io/}
}
