% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotion.R
\name{emotion}
\alias{emotion}
\title{R interface to tencent AI lab's emotion analysis}
\usage{
emotion(text, app_key, app_id)
}
\arguments{
\item{text}{The text is sent to API}

\item{app_key}{Your app key}

\item{app_id}{Your app ID}
}
\value{
positive = 1, neuter = 0, nagetive = -1
}
\description{
\code{emotion} returns the emotion analysis APP results
}
\author{
Ao Sun <\url{https://ao-sun.github.io/}
}
