% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{R interface to tencent AI text translation}
\usage{
translate(app_id = app_id, app_key = app_key, keywords)
}
\arguments{
\item{app_id}{Your app ID}

\item{app_key}{Your app key}

\item{text}{The text is sent to API}
}
\description{
This function can only handle English and Chinese
}
\details{
\code{translate} returns the result of transform
}
\examples{
 translate('god', app_id = '1107152791', app_key =  'OpsMj8HXPmbu9SMd')
}
\author{
Ao Sun <\url{https://ao-sun.github.io/}>
}
