% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotion.R
\name{emotion}
\alias{emotion}
\title{R interface to tencent AI lab's emotion analysis}
\usage{
emotion(text, app_key, app_id)
}
\arguments{
\item{text}{The text is sent to API}

\item{app_key}{Your app key}

\item{app_id}{Your app ID}
}
\value{
positive = 1, neuter = 0, nagetive = -1
}
\description{
text can only be Chinese, this function performs poorly in English
}
\details{
\code{emotion} returns the results  of emotion analysis
}
\examples{
 emotion('很开心', app_id = '1107152791', app_key =  'OpsMj8HXPmbu9SMd')
}
\author{
Ao Sun <\url{https://ao-sun.github.io/}>
}
